+++ spacypdfreader/__init__.py
@@ -1,5 +1,3 @@
 import importlib.metadata
 
-from spacypdfreader.spacypdfreader import pdf_reader
-
-__version__ =  importlib.metadata.version("spacypdfreader")
+__version__ = importlib.metadata.version("spacypdfreader")